<STYLE>: norm
