+++ srcs/ft_export3.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_export3.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/27 21:25:38 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/07/03 20:31:12 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/07/07 18:08:12 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,6 @@ void	sort_list(t_list *env, t_list **top)
 	t_env = t_env->next;
 	while (t_env)
 	{
-		// printf("%s, %s\n", t_env->key, t_env->value);
 		sorting(t_env, top);
 		t_env = t_env->next;
 	}
