+++ srcs/get_fd2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_fd2.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/31 21:22:40 by heerpark          #+#    #+#             */
-/*   Updated: 2024/07/03 20:10:30 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/07/07 18:04:50 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,6 +53,7 @@ int	get_heredoc(t_head *head, t_process *process, char *limiter)
 int	get_outfile(t_head *head, t_process *process, char *file_name)
 {
 	int	fd;
+
 	fd = open(file_name, O_RDWR | O_CREAT | O_TRUNC, 0644);
 	if (fd == -1)
 	{
