+++ srcs/list_to_process_utils1.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   list_to_process_utils1.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/31 21:53:55 by heerpark          #+#    #+#             */
-/*   Updated: 2024/07/03 20:10:38 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/07/07 17:46:54 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,9 @@ void	init_process(t_process *process)
 	process->is_error = 0;
 }
 
-int	get_redir_flag(char	*token)
+int	get_redir_flag(char	*token, int *is_filename)
 {
+	*is_filename = 1;
 	if (ft_strlen(token) == 1)
 	{
 		if (ft_strncmp(token, ">", 2) == 0)
@@ -82,7 +83,6 @@ void	close_unused_output(t_process *process)
 void	set_fd(t_head *head, t_process *process, \
 		char *file_name, int redir_flag)
 {
-	// printf("redir flag %d\n", redir_flag);
 	if (redir_flag == 1)
 	{
 		close_unused_output(process);
