+++ srcs/main.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/16 17:10:35 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/07/03 16:13:59 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/07/07 18:05:53 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/minishell.h"
 
-int g_exit_status;
+int	g_exit_status;
 
 int	check_white_space(char *str)
 {
@@ -45,8 +45,6 @@ int	main(int argc, char **argv, char **envp)
 	char	*str;
 	t_head	*head;
 
-	// if (ft_strcmp(">>", ">>>") == 0)
-	// 	printf("hi\n");
 	head = init_head(envp, argc, argv);
 	while (1)
 	{
@@ -63,8 +61,6 @@ int	main(int argc, char **argv, char **envp)
 				parse_error(str, head);
 				continue ;
 			}
-			// system("leaks minishell");
-			// sleep(2);
 			update_envp(head);
 			exe(head, head->data->envp);
 			clear(head, str);
