+++ srcs/processes_exe.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   processes_exe.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: heerpark <heerpark@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/24 03:11:25 by heerpark          #+#    #+#             */
-/*   Updated: 2024/07/03 20:24:11 by heerpark         ###   ########.fr       */
+/*   Updated: 2024/07/07 18:01:20 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,9 +37,7 @@ void	run_cmd(t_head *head, char **envp, int i)
 	if (head->processes[i]->is_error)
 	{
 		if (head->processes[i]->is_error == 42)
-		{
 			exit(0);
-		}
 		exit(head->processes[i]->is_error);
 	}
 	if (is_builtin(head->processes[i]->exec_cmd))
@@ -129,11 +127,8 @@ void	exe(t_head *head, char **envp)
 	{
 		head->data->pipes = make_pipe(head->size - 1);
 		get_processes(head, envp);
-		// if (error_check(head, 1))
-		// 	return ;
 		start_processes(head, envp, head->data->pipes);
 		wait_process(head->size, head->data->last_pid);
 	}
-
 	set_signal();
 }
