+++ srcs/split_space.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   split_space.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/21 22:41:50 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/07/03 20:16:57 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/07/07 18:07:28 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,57 +69,8 @@ char	*make_cmd(char *cmd, t_split_var *sv, char q)
 	}
 	p_cmd[v.ind] = 0;
 	sv->i = v.i;
-	// sv->start = v.i;
-	// printf("cmd : %s\n", p_cmd);
-	// printf("cmd[i] : %c\n", cmd[v.i]);
 	return (p_cmd);
 }
-
-char	*ft_strjoin2(char const *s1, char const *s2)
-{
-	size_t	len;
-	size_t	i;
-	char	*str;
-
-	if (!s1 || !s2)
-		return (NULL);
-	i = 0;
-	len = ft_strlen((char *)s1) + ft_strlen((char *)s2);
-	str = (char *)malloc(sizeof(char) * (len + 1));
-	if (!str)
-		return (NULL);
-	while (*s1)
-		str[i++] = *s1++;
-	while (*s2)
-		str[i++] = *s2++;
-	str[i] = 0;
-	return ((char *)str);
-}
-
-// void	split_space_ext(t_split_var *v, char *cmd)
-// {
-// 	char	*tmp;
-
-// 	if (cmd[v->i] == '\0')
-// 		v->flag = 1;
-// 	cmd[v->i] = '\0';
-// 	if (v->i > v->start && v->backup)
-// 	{
-// 		tmp = v->backup;
-// 		v->backup = ft_strjoin2(tmp, &cmd[v->start]);
-// 		if (!v->backup)
-// 			error_msg(0, NULL);
-// 		free(tmp);
-// 	}
-// 	if (!v->backup)
-// 		add_token(&v->lst, &cmd[v->start]);
-// 	else
-// 		ft_token_add(&v->lst, token_new(ft_strdup(v->backup), 0, v->quote));
-// 	free(v->backup);
-// 	v->backup = NULL;
-// 	v->i++;
-// 	v->start = v->i;
-// }
 
 int	is_quote(char c)
 {
@@ -128,69 +79,19 @@ int	is_quote(char c)
 	return (0);
 }
 
-// void	if_redir_exist(t_split_var *v, char *cmd)
-// {
-	
-// }
-
-void	split_space_ext(t_split_var *v, char *cmd)
-{
-	// int	i;
-
-	// i = 0;
-	printf("origin backup: %s\n", v->backup);
-	printf("----------------cmd[v->i] %c  %d\n",cmd[v->i - v->start], v->i);
-	if (cmd[v->i - v->start] == '\0')
-		v->flag = 1;
-	cmd[v->i - v->start] = '\0';
-	// if_redir_exist()
-	// while (cmd[i] && !is_redir(cmd[i]) && !is_quote(cmd[i]))
-	// 	i++;
-	// while (cmd[i] && !is_redir(cmd[i]) && !is_quote(cmd[i]))
-	// v->start += i;
-	// add_token(&v->lst, ft_strndup(cmd, i));
-	// cmd = cmd + i;
-	// if (v->i > v->start && v->backup)	// backup이 있었다면 둘이합치기
-	// {
-	// 	tmp = v->backup;
-	// 	v->backup = ft_strjoin(v->backup, &cmd[v->start]);
-	// 	free(tmp);
-	// 	printf("backup: %s\n", v->backup);
-	// }
-	if (!v->backup)
-	{
-		printf("h1\n");
-		add_token(&v->lst, ft_strdup(cmd));
-	}
-	else
-	{
-		printf("h2\n");
-		add_token2(&v->lst, ft_strdup(v->backup), v->quote);
-		// v->lst = ft_token_add2(v->lst, token_new(ft_strdup(v->backup), 0, v->quote));
-		free(v->backup);
-	}
-	v->i++;
-	v->start = v->i;
-	v->backup = NULL;
-	free(cmd);
-	printf("after: %s\n", v->lst->cmd);
-}
-
-
 t_token	*split_space(char *cmd, char space)
 {
 	t_split_var	v;
-	// char		*tmp;
 
 	v = (t_split_var){NULL, NULL, 0, 0, 0, 0, 0};
 	v.dquote = 0;
 	v.quote = 0;
 	while (v.flag == 0 && cmd[v.i])
 	{
-		while (cmd[v.start] == space)	// 맨앞 공백 넘김
+		while (cmd[v.start] == space)
 			v.start++;
 		v.i = v.start;
-		while ((v.dquote || v.quote || cmd[v.i] != space) && cmd[v.i])	// space 전까지 계속 넘기는데, quote 내부라면 space만나도 계속 넘길 수 있음.
+		while ((v.dquote || v.quote || cmd[v.i] != space) && cmd[v.i])
 		{
 			my_quote_check(cmd[v.i], &v.quote, &v.dquote);
 			v.i++;
@@ -198,14 +99,5 @@ t_token	*split_space(char *cmd, char space)
 		add_token(&v.lst, ft_strndup(&cmd[v.start], v.i - v.start));
 		v.start = v.i;
 	}
-
-	// t_token *a = v.lst;
-	// while (a)
-	// {
-	// 	printf(".... %s %d %d\n", a->cmd, a->redir_flag, a->quote_flag);
-	// 	a = a->next;
-	// }
-	// printf(".... |\n");
-
 	return (v.lst);
 }
