+++ srcs/utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhyeong <junhyeong@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 18:17:19 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/07/03 17:37:00 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/07/07 18:07:59 by junhyeong        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,7 +75,7 @@ void	free_list(t_head *head, char *str)
 	}
 	if (head->processes)
 		clear_processes(head);
-	if (head->error_str)		
+	if (head->error_str)
 		free(head->error_str);
 	head->error_str = NULL;
 	head->size = 0;
